# specs/002-fastapi-dashboard-api/contracts/openapi-spec.yaml
openapi: 3.0.3
info:
  title: RiseTrader Dashboard API
  description: |
    FastAPI service serving market data, forecasts, account information, and strategy allocations
    from PostgreSQL database to React dashboard with real-time MT4 integration.

    ## Real-Time Updates
    The API provides Server-Sent Events (SSE) endpoints for real-time data streaming:
    - `/api/stream/market-data` - Price updates
    - `/api/stream/account` - Account balance and positions
    - `/api/stream/forecasts` - New forecast availability

    ## Authentication
    ⚠️ Currently NO authentication implemented. API key + JWT authentication will be added in Phase 8-9.

    ## Rate Limiting
    Rate limiting is enforced via slowapi middleware. Default limits apply per IP address.

    ## Pagination
    List endpoints use keyset (cursor-based) pagination for consistent performance:
    - `page`: Page number (1-indexed)
    - `page_size`: Items per page (default: 50, max: 500)
    - `cursor`: Opaque cursor for next page (returned in `next_cursor` field)
  version: 1.0.0
  contact:
    name: RiseTrader Team
    email: support@risetrader.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:8003/api
    description: Local development server
  - url: https://api.risetrader.com/api
    description: Production server (future)
...
paths:
...
  /trading/positions:
    get:
      tags:
        - trading
      summary: Get open positions
      description: |
        Retrieve all currently open trading positions.

        **Real-time**: Updates within 3 seconds of MT4 changes (success criteria SC-003)
      operationId: getPositions
      parameters:
        - name: symbol
          in: query
          required: false
          description: Filter by symbol
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PositionListResponse'
  /trading/history:
    get:
      tags:
        - trading
      summary: Get trading history
      description: |
        Retrieve completed trading transactions with pagination.

        **Data Retention**: 90 days minimum (success criteria SC-007)
      operationId: getTradingHistory
      parameters:
        - name: symbol
          in: query
          required: false
          schema:
            type: string
        - name: start_date
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: trade_type
          in: query
          required: false
          schema:
            type: string
            enum: [BUY, SELL]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradeListResponse'

  # ==========================================
  # FORECASTS ENDPOINTS
  # ==========================================
...
components:
  schemas:
...
    PositionResponse:
      type: object
      properties:
        id:
          type: integer
        ticket:
          type: integer
          example: 123456
        symbol:
          type: string
          example: "CrudeOIL"
        type:
          type: string
          enum: [BUY, SELL]
        lots:
          type: number
          format: double
          example: 0.1
        open_price:
          type: number
          format: double
          example: 75.00
        current_price:
          type: number
          format: double
          example: 75.45
        profit:
          type: number
          format: double
          example: 45.00
        open_time:
          type: string
          format: date-time
        stop_loss:
          type: number
          format: double
          nullable: true
        take_profit:
          type: number
          format: double
          nullable: true
    PositionListResponse:
      type: object
      properties:
        positions:
          type: array
          items:
            $ref: '#/components/schemas/PositionResponse'
        total:
          type: integer
    TradeResponse:
      type: object
      properties:
        id:
          type: integer
        ticket:
          type: integer
        symbol:
          type: string
        type:
          type: string
          enum: [BUY, SELL]
        lots:
          type: number
          format: double
        entry_price:
          type: number
          format: double
        exit_price:
          type: number
          format: double
        entry_time:
          type: string
          format: date-time
        exit_time:
          type: string
          format: date-time
        profit:
          type: number
          format: double
        commission:
          type: number
          format: double
          nullable: true
        swap:
          type: number
          format: double
          nullable: true
        duration_seconds:
          type: integer
    TradeListResponse:
      type: object
      properties:
        trades:
          type: array
          items:
            $ref: '#/components/schemas/TradeResponse'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer

    # ==========================================
    # FORECAST SCHEMAS
    # ==========================================
...